Fix multithreading
The arrays specifying the requested image block were shared by all threads,
such that they were accessing the same image block and updating the ROI
of the next requested block in the shared arrays (for all threads).
This created frequent segfaults and/or a repetitive image. All required
objects are now local.